requirements.txt
streamlit==1.41.1

langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-google-genai==1.0.10
google-generativeai==0.7.2

langgraph==0.2.33


PyPDF2==3.0.1
faiss-cpu==1.8.0.post1
python-docx==1.1.2
pandas==2.2.3
typing_extensions==4.12.2
